MOD: change loss function to grid method
The previous use of the probability model-based loss function significantly reduced
the probability of a stuck at the usual frequencies, but the total number of lags
did not decrease, causing lags to converge to higher frequency segments, which
reduced the response speed. The grid-count based loss function attempts to solve
this problem, in conjunction with a load model that contains multiple scenarios,
focusing on worst-case and mainstream situations to improve overall performance.